requirements.txt
streamlit==1.43.2
agno==1.1.17
python-dotenv==1.0.1